score -> pvalue
