+++ cointanalysis/coint.py
@@ -2,7 +2,7 @@ import numpy as np
 from sklearn.base import TransformerMixin
 from sklearn.decomposition import PCA
 from sklearn.linear_model import LinearRegression
-from statsmodels.tsa.stattools import adfuller, coint
+from statsmodels.tsa.stattools import coint
 
 
 class CointAnalysis(TransformerMixin):
@@ -71,7 +71,7 @@ class CointAnalysis(TransformerMixin):
         self.__check_trend(trend)
         self.trend = trend
 
-    def score(self, X, y=None, method='AEG', stat_method='ADF', stat_pvalue=.05):
+    def pvalue(self, X, y=None, method='AEG', stat_method='ADF', stat_pvalue=.05):
         """
         Return p-value of cointegration test.
         Null-hypothesis is no cointegration.
@@ -226,61 +226,3 @@ class CointAnalysis(TransformerMixin):
     def __rms(array):
         """Root-mean square"""
         return np.sqrt(array.mean() ** 2 + array.std() ** 2)
-
-
-class StationarityTest:
-    """
-    Parameters
-    ----------
-    - method : {'ADF'}, default 'ADF'
-        Test method.
-    - regression : {'c', 'ct', 'ctt', 'nc'}
-        Regression method.
-
-    Examples
-    --------
-    >>> np.random.seed(42)
-    >>> gauss = np.random.randn(100)  # stationary
-    >>> stat = StationarityAnalisis()
-    >>> stat.score(gauss)  # returns p-value
-    1.16e-17
-    >>> brown = gauss.cumsum()
-    >>> stat.score(brown)  # returns p-value
-    0.60
-    """
-    def __check_method(self, method):
-        __available = ('ADF', )
-        if method not in __available:
-            raise ValueError(f'Invalid method: {method}')
-
-    def __check_regression(self, regression):
-        __available = ('c', 'ct', 'ctt', 'nc')
-        if regression not in __available:
-            raise ValueError(f'Invalid regression: {regression}')
-
-    def __init__(self, method='ADF', regression='c'):
-        """Initialize self."""
-        self.__check_method(method)
-        self.__check_regression(regression)
-
-        self.method = method
-        self.regression = regression
-
-    def score(self, X, y=None):
-        """
-        Return p-value of unit-root test.
-        Null-hypothesis is that there is a unit root (not stationary).
-
-        Parameters
-        ----------
-        - X : array, shape (n_samples, 2)
-            A pair of price data, where n_samples is the number of samples.
-        - y : None
-            Ignored.
-        """
-        if self.method == 'ADF':
-            _, pvalue, _, _, _, _ = adfuller(X, regression=self.regression)
-            return pvalue
-
-    def is_stationary(self, X, y=None, pvalue=.05):
-        return self.score(X) < pvalue
